* master: (108 commits)
  version info: update
  Common.h: Fix compilation warnings
  Putty: Use dlmalloc
  WebDAV: Fix memory management
  neon: Fix memory management
  FileZillaIntf.cpp
  TStringItemList
  UnicodeString.cpp: Use wmemmove
  Optimization
  CopyParam.cpp
  custom_nballocator_t
  Revert "Optimization"
  custom_nballocator_t
  Optimization
  RemoteFiles.cpp: Optimization
  FtpFileSystem.cpp: Remove TRACING
  Common.cpp: Remove TRACING
  Sysutils.cpp: vswprintf
  headers.hpp: custom_nballocator_t
  Common.cpp: Remove TRACING
  ...

Conflicts:
	src/NetBox/CMakeLists.txt
	src/NetBox/FarDialog.cpp
	src/NetBox/FarPlugin.cpp
	src/NetBox/NetBox.vcxproj
	src/NetBox/scripts/build_netbox.cmd
	src/NetBox/scripts/make_dist.cmd

+++ src/NetBox/WinSCPPlugin.cpp
@@ -651,8 +651,8 @@ void TWinSCPPlugin::MessageClick(void * Token, intptr_t Result, bool & Close)
   }
 }
 //---------------------------------------------------------------------------
-intptr_t TWinSCPPlugin::MoreMessageDialog(const UnicodeString & Str,
-  TStrings * MoreMessages, TQueryType Type, int Answers,
+uintptr_t TWinSCPPlugin::MoreMessageDialog(const UnicodeString & Str,
+  TStrings * MoreMessages, TQueryType Type, uintptr_t Answers,
   const TMessageParams * Params)
 {
   intptr_t Result = 0;
@@ -661,7 +661,7 @@ intptr_t TWinSCPPlugin::MoreMessageDialog(const UnicodeString & Str,
   {
     std::auto_ptr<TStrings> ButtonLabelsPtr;
     ButtonLabelsPtr.reset(ButtonLabels);
-    unsigned int Flags = 0;
+    uintptr_t Flags = 0;
 
     if (Params != NULL)
     {
@@ -703,7 +703,7 @@ intptr_t TWinSCPPlugin::MoreMessageDialog(const UnicodeString & Str,
       }
     }
 
-    int AAnswers = Answers;
+    uintptr_t AAnswers = Answers;
     bool NeverAskAgainCheck = (Params != NULL) && FLAGSET(Params->Params, qpNeverAskAgainCheck);
     bool NeverAskAgainPending = NeverAskAgainCheck;
     intptr_t TimeoutButton = 0;
@@ -779,14 +779,14 @@ intptr_t TWinSCPPlugin::MoreMessageDialog(const UnicodeString & Str,
     {
       if (Params->Timer > 0)
       {
-        FarParams.Timer = Params->Timer;
+        FarParams.Timer = static_cast<unsigned int>(Params->Timer);
         FarParams.TimerEvent = Params->TimerEvent;
       }
 
       if (Params->Timeout > 0)
       {
         FarParams.Timeout = Params->Timeout;
-        FarParams.TimeoutButton = (unsigned int)TimeoutButton;
+        FarParams.TimeoutButton = static_cast<unsigned int>(TimeoutButton);
         FarParams.TimeoutStr = GetMsg(MSG_BUTTON_TIMEOUT);
       }
     }
@@ -803,7 +803,7 @@ intptr_t TWinSCPPlugin::MoreMessageDialog(const UnicodeString & Str,
       FarParams.MoreMessages = NULL;
     }
 
-    Result = Message(Flags, GetMsg(TitleId), DialogStr, ButtonLabels, &FarParams);
+    Result = Message(static_cast<DWORD>(Flags), GetMsg(TitleId), DialogStr, ButtonLabels, &FarParams);
     if (FarParams.TimerAnswer > 0)
     {
       Result = static_cast<intptr_t>(FarParams.TimerAnswer);
